pyproject.toml
[project]
name = "ml-intern"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "enricher>=0.1.0",
    "fastapi>=0.118.0",
    "groq>=0.32.0",
    "openai>=2.2.0",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.14.1",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "uvicorn>=0.37.0",
]
